Make mock import python 2.7 compatible
In python3.x mock is now part of unittest library.
For being still compatible with python 2.7 we need to
catch import error when trying to import unittest.mock.

Also add an __init__.py file in validations_common/library,
to allow all unit tests to import successfully.

This is similar to [1] for validations-libs.

[1]- https://review.opendev.org/716597

Change-Id: I6a81dc2640e9c4af3a07158bea95ebaec9fd466b