Move unit tests execution to stestr
From time to time, unit tests execution fails with error:
"TypeError: 'NoneType' object is not iterable". It seems that
problem is in unittest2 library, which is used by testtools for
executing the tests.
Stestr lib is usinig unittest by default and the problem does not appear
for several hundred launches.

Change-Id: Ic1053353a87ebd44f20c607651772928020b9e66
Closes-Bug: #1874535