Add validation cli script
In order to be able to run validation without any
TripleO package installed and as a first step before
an Undercloud deployment, this review add a simple
CLI implementation which expose basic commands in order
to run, list and show validation on a system.

The user will only have to install either:
tripleo-validations rpm package or
pip install validations-common and validations-libs.

Change-Id: I7131ba18b625da282a529473cd804dbdae5a424e