Remove RoleNames ansible-lint custom rule
As this rule has been merged upstream, we don't need it anymore locally.

This patch also bumps the ansible-lint release from v4.2.0 to v5.0.3

It also mocks the following roles to pass the --syntax-check:
- fetch-devstack-log-dir
- fetch-output
- fetch-python-sdist-output
- fetch-subunit-output
- ensure-if-python
- ensure-pip
- ensure-tox
- ensure-virtualenv

Finally, it deletes two empty molecule verify.yml playbooks which are
causing the linter to fail.

Change-Id: I132e2144835724ec82eb0a9aee1996d1e1c3e5fb
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>