Introduce tox-ansible to validations-common
This patch adds basic configuration for tox-ansible[1].

Tox-ansible is a plugin for tox which auto-generates tox environments
for running quality assurance tools like ansible-test or molecule. This
tool is rather tightly integrated for the official Molecule testing tool
that integrates with Ansible.

At first, it will allow us to list very quickly all the Ansible Molecule
tests for each roles with their scenario(s):

Example:

    $ tox -l
    ...
    validations_common-roles-advanced_format_512e_support-default
    validations_common-roles-check_cpu-default
    validations_common-roles-check_disk_space-default
    validations_common-roles-check_latest_packages_version-default
    validations_common-roles-check_ram-default
    validations_common-roles-check_selinux_mode-default
    validations_common-roles-dns-default
    validations_common-roles-haproxy-default
    validations_common-roles-ntp-default
    validations_common-roles-service_status-default
    validations_common-roles-service_status-docker
    validations_common-roles-service_status-podman
    validations_common-roles-service_status-systemd
    validations_common-roles-validate_selinux-default
    validations_common-roles-xfs_check_ftype-default
    ...

However, we won't be able to execute them as it for now without to make
some small manual modifications in the molecule.yml file first before to
execute them. This will be covered as soon as possible in a following-patch.

[1] - https://github.com/ansible-community/tox-ansible

Change-Id: Ia876bb1634f98e378345939788280e71738d03c2
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>