Fix targeting hosts to 'localhost' instead of 'all'
This patch fixes the targeting hosts for the validations which should be
executed locally.

Rationale: When those validations are executed in the TripleO world,
they are executed on all the nodes including the Overcloud nodes. For
instance, check-ram and check-cpu validations check the Red Hat
recommendations for the Undercloud only and shouldn't be checking the
Overcloud nodes.

Change-Id: Icca554ea4392c3b0b572b27c2e2370b11f8013d7
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>