File validation extended to callback_plugins, main loop optimized
validate-files.py now also checks required attributes
of callback plugins. At this time, only two commonly used
attributes are checked, because not all of our plugins
correctly declare type and documentation.

Main loop of the script was adjusted to work with multiple
types of files to validate, and to minimize unnecessary checks.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I2c92b77f127a82afaf28a3dd5ebfb9e3614bbb07