Removing the 'PIP_+' string from tox.ini
Presence of the string causes following error,
when running molecule tests on local:
'Error: error parsing host environment variables: name "PIP_+  VIRTUAL_ENV" has white spaces, poorly formatted name'

After removing it the molecule tests are able to run successfuly.

Furthermore I found no reference to the purpose of the string
in either the tox documentation, or by searching for the same
string on github. It appears we are the only project hosting code
on the service, using this particular setting.

I therefore assume that it is nonessential.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I61641a1d4ddd017451b4aa7aea1e6eae6439b53d