[ntp] Add role documentation in sphinx
This patch adds the documentation directly in the sphinx role
documentation index. All variables have been commented with a proper
description to be included automatically in the sphinx build.

This patch also disable the facts gathering because this validation
doesn't need them and only uses service_facts in the ntp role.

Change-Id: I3a19eb0a93f6dfaaf7a4361b46b2dc904aeb7ae6
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>