[validate_selinux] Add role documentation in sphinx
This patch adds the documentation directly in the sphinx role
documentation index. All variables have been commented with a proper
description to be included automatically in the sphinx build.

This patch changes the default value for the
validate_selinux_working_dir variable to /tmp instead of
/var/log/validations to be consistent with the default used in the
playbook.

Moreover, this validation will warn the user when Untracked SELinux AVCs
are found instead of using the debug module and the strict mode
information is now displayed at the end.

NB: this patch also explicitly disables the facts gathering because this
validation doesn't need them.

Change-Id: Id337cbd89ab9368ee7e02d9c710861b134929ec7
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>