callback adjustment for no hosts matched
The validation_json callback will now consider the no hosts being matched
case as a validation failure. And record the result as such.

To prevent unnecessary confusion, the host indicated in the created
log file will be designated as 'No host matched'. And will be displayed
as such on the VF stdout.

Basic test coverage included.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I613203698b726941162239185de77949bf30254c