Test validations for false positives
By running validations with arguments intentionally chosen
to make the validation requirements impossible to fulfill,
we can check how the framework handles failures, and if
the tested validations do not produce false positives.

The functional test now, in addition to the existing
set of checks, allow for optional 'negative_results' testing.

Variables entries for all tested validations are checked
for the presence of 'negative_results' dictionary key.
If it is found, the 'extra_args' key contents within are used
to run the validation in question, after it has passed the test
for a successful validation run.

The 'negative results' run is expected to fail, failure is supressed,
and result recorded. Passing result, conversely, is interpreted as an error,
because validation passing shouldn't be possible with the arguments supplied.

Show validations test was patched to proper form of the command.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: If7e7b25b9c876184bd04f3f0ddefdee20172a5b4