CI task for execution of all packaged validations
Task 'execute_full_catalogue' invokes validation command
with list of validations constituting entirety of the existing catalogue.

Validations that are to be executed are listed in the variables
fail of the 'validations' role.

Removal of the validation from the list has to be accompanied by
placement of the aformentioned validation in the commented form
beneath the list. With a concise reasoning as to why was it removed.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Iea16cdc46db53cb290f7d31436864436fd31dd40