Run all validations and fail after for CI role
The execution of validations in the CI role was checking the
result on each run and fail is something was not expected.
This patch failed the execution only after all the validations has been run.

This allow better analysis and debuging when more than one validations
are failing on a job.

Change-Id: I9b8f825a02f412475377cee34d88abe8431bc318