Setting default branch of devstack repositories for functional tests
The functional test job didn't specify default branches
for repositories it depended on, leading to a graph freeze failure
on stable branches with irregular names.

This patch sets the devstack repos as a dependency and
explicitly specifies the branches to be used.

Closes-Bug: #1965876
Co-Authored-By: Jakob Meng <code@jakobmeng.de>

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Id22cd56416849da03d02ef98781b601f223355db
(cherry picked from commit a679bb8015437d0d5cec65223549d8535291823f)