Setting default inventory to 0 length string
Default inventory used by the CI has to be set to
a string of zero length in order to trigger inventory discovery
mechanism, otherwise the task is skipped.

The default inventory "localhost" is substituted only if no viable path is discovered.

Loop variables of the inventory discovery task were explicitly named and conditionals simplified
in order to improve overall stability of the routine.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Icccc5384cfc5d04e8c92682e7cbab5acad5a4e39