[Stable/1.6-Only] Setting docs job and bindep for stable/1.6
Patch sets branch override on the doc building job and constrains
the sphinx version used for doc building

libffi headers package was added in bindep, in order to eliminate build failure of the cffi package.

The previous minimum version of Sphinx allowed for incompatible extensions to be installed,
in certain edge cases. Incrementing and capping the minimum version allowed addresses the issue.

Related-Bug: #1980465

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I027ba43989e75c35de01400356acfc9d36fc1d24