Set fact for run_results returns error when previous validation run is skipped
When validation run is skipped due the unfulfilled condition, run_results variable isn't created, so the set_fact cannot be set properly and returns error which can be interpreted as false negatives. Set_fact task must be corrected so it doesn't fail even if the variable run_results is empty.

Signed-off-by: Veronika Fisarova <vfisarov@redhat.com>
Change-Id: Ibb6bf62c0291e27da6b855d7836376a029233138